requirements_dev.txt
freezegun
django_extensions
pytest
pytest-django
ipython
ipdb
mock
vcrpy